requirements.txt
captcha==0.3
numpy==1.18.2
protobuf==3.11.3
spacy==2.2.4
torch==1.4.0
torchvision==0.5.0
torchaudio==1.5.0
websockets==8.1
scipy==1.3.3
ffmpeg-python==0.2.0
psutil==5.7.0
pathos==0.2.5
tensorflow-gpu==1.10.0
keras==2.2.2
matplotlib==2.2.2